Harden incremental deployment correlation
