+++ .github/workflows/powerforge-website-deploy.yml
@@ -268,6 +268,7 @@ jobs:
     outputs:
       run_id: ${{ steps.deployment_provenance.outputs.run_id }}
       run_attempt: ${{ steps.deployment_provenance.outputs.run_attempt }}
+      job_id: ${{ steps.deployment_provenance.outputs.job_id }}
     environment:
       name: github-pages
       url: ${{ steps.deployment.outputs.page_url || steps.deployment_retry.outputs.page_url }}
@@ -294,6 +295,7 @@ jobs:
         run: |
           "run_id=${{ github.run_id }}" | Out-File -FilePath $env:GITHUB_OUTPUT -Encoding utf8 -Append
           "run_attempt=${{ github.run_attempt }}" | Out-File -FilePath $env:GITHUB_OUTPUT -Encoding utf8 -Append
+          "job_id=${{ job.check_run_id }}" | Out-File -FilePath $env:GITHUB_OUTPUT -Encoding utf8 -Append
 
   cloudflare-site-policy:
     if: ${{ inputs.deployment_target == 'github-pages' && inputs.manage_cloudflare_site_policy }}
@@ -332,6 +334,7 @@ jobs:
           deployment-id: ${{ needs.build.outputs.source_sha }}
           deployment-run-id: ${{ needs.deploy.outputs.run_id }}
           deployment-run-attempt: ${{ needs.deploy.outputs.run_attempt }}
+          deployment-job-id: ${{ needs.deploy.outputs.job_id }}
           manifest-artifact-name: ${{ needs.build.outputs.cloudflare_cache_manifest_artifact_name }}
           github-token: ${{ github.token }}
