+++ PowerForge.Tests/CloudflareIncrementalCachePurgeTests.cs
@@ -10,7 +10,7 @@ using PowerForge.Web.Cli;
 
 namespace PowerForge.Tests;
 
-public sealed class CloudflareIncrementalCachePurgeTests
+public sealed partial class CloudflareIncrementalCachePurgeTests
 {
     private const string ZoneId = "0123456789abcdef0123456789abcdef";
 
@@ -70,6 +70,87 @@ public sealed class CloudflareIncrementalCachePurgeTests
                 Path.Combine(root, "manifest.json")));
 
             Assert.Contains("unsupported type", exception.Message, StringComparison.OrdinalIgnoreCase);
+        }
+        finally
+        {
+            Directory.Delete(root, recursive: true);
+        }
+    }
+
+    [Fact]
+    public void CreateManifest_ShouldResolveSafeInArchiveHardLinks()
+    {
+        var root = NewTempDirectory();
+        try
+        {
+            var artifactPath = Path.Combine(root, "artifact.tar");
+            using (var stream = File.Create(artifactPath))
+            using (var writer = new TarWriter(stream, TarEntryFormat.Pax, leaveOpen: false))
+            {
+                writer.WriteEntry(new PaxTarEntry(TarEntryType.RegularFile, "assets/original.css")
+                {
+                    DataStream = new MemoryStream(Encoding.UTF8.GetBytes("body{}"), writable: false)
+                });
+                writer.WriteEntry(new PaxTarEntry(TarEntryType.HardLink, "assets/copy.css")
+                {
+                    LinkName = "assets/original.css"
+                });
+                writer.WriteEntry(new PaxTarEntry(TarEntryType.HardLink, "assets/copy-chain.css")
+                {
+                    LinkName = "assets/copy.css"
+                });
+            }
+
+            var result = CloudflareDeploymentManifestStore.CreateFromTar(
+                artifactPath,
+                "https://example.test/",
+                Path.Combine(root, "manifest.json"));
+            var manifest = CloudflareDeploymentManifestStore.LoadRequired(result.ManifestPath);
+
+            Assert.Equal(3, result.ArtifactFileCount);
+            Assert.Equal(18, result.ContentBytes);
+            Assert.Equal(3, manifest.Files.Length);
+            Assert.Single(manifest.Files.Select(entry => entry.Sha256).Distinct(StringComparer.Ordinal));
+            Assert.All(manifest.Files, entry => Assert.Equal(6, entry.Length));
+        }
+        finally
+        {
+            Directory.Delete(root, recursive: true);
+        }
+    }
+
+    [Fact]
+    public void CreateManifest_ShouldRejectUnsafeOrUnresolvedHardLinks()
+    {
+        var root = NewTempDirectory();
+        try
+        {
+            foreach (var target in new[] { "../outside.css", "missing.css" })
+            {
+                var artifactPath = Path.Combine(root, $"{Guid.NewGuid():N}.tar");
+                using (var stream = File.Create(artifactPath))
+                using (var writer = new TarWriter(stream, TarEntryFormat.Pax, leaveOpen: false))
+                    writer.WriteEntry(new PaxTarEntry(TarEntryType.HardLink, "copy.css") { LinkName = target });
+
+                var exception = Assert.Throws<InvalidDataException>(() => CloudflareDeploymentManifestStore.CreateFromTar(
+                    artifactPath,
+                    "https://example.test/",
+                    Path.Combine(root, $"{Guid.NewGuid():N}.json")));
+                Assert.Contains(target.StartsWith("..", StringComparison.Ordinal) ? "unsafe" : "unavailable", exception.Message, StringComparison.OrdinalIgnoreCase);
+            }
+
+            var cyclePath = Path.Combine(root, "cycle.tar");
+            using (var stream = File.Create(cyclePath))
+            using (var writer = new TarWriter(stream, TarEntryFormat.Pax, leaveOpen: false))
+            {
+                writer.WriteEntry(new PaxTarEntry(TarEntryType.HardLink, "a.css") { LinkName = "b.css" });
+                writer.WriteEntry(new PaxTarEntry(TarEntryType.HardLink, "b.css") { LinkName = "a.css" });
+            }
+            var cycleException = Assert.Throws<InvalidDataException>(() => CloudflareDeploymentManifestStore.CreateFromTar(
+                cyclePath,
+                "https://example.test/",
+                Path.Combine(root, "cycle.json")));
+            Assert.Contains("cycle", cycleException.Message, StringComparison.OrdinalIgnoreCase);
         }
         finally
         {
@@ -525,6 +606,9 @@ public sealed class CloudflareIncrementalCachePurgeTests
         Assert.Contains("manifest-artifact-name: ${{ needs.build.outputs.cloudflare_cache_manifest_artifact_name }}", deployWorkflow, StringComparison.Ordinal);
         Assert.Contains("deployment-run-id: ${{ needs.deploy.outputs.run_id }}", deployWorkflow, StringComparison.Ordinal);
         Assert.Contains("deployment-run-attempt: ${{ needs.deploy.outputs.run_attempt }}", deployWorkflow, StringComparison.Ordinal);
+        Assert.Contains("job_id: ${{ steps.deployment_provenance.outputs.job_id }}", deployWorkflow, StringComparison.Ordinal);
+        Assert.Contains("job_id=${{ job.check_run_id }}", deployWorkflow, StringComparison.Ordinal);
+        Assert.Contains("deployment-job-id: ${{ needs.deploy.outputs.job_id }}", deployWorkflow, StringComparison.Ordinal);
         Assert.Contains("actions: write", deployWorkflow, StringComparison.Ordinal);
         Assert.Contains("deployments: read", deployWorkflow, StringComparison.Ordinal);
         Assert.Contains("manage-incremental-purge: \"true\"", deployWorkflow, StringComparison.Ordinal);
@@ -532,6 +616,8 @@ public sealed class CloudflareIncrementalCachePurgeTests
         Assert.Contains("github-token: ${{ github.token }}", deployWorkflow, StringComparison.Ordinal);
         Assert.Contains("Resolve-PowerForgeRepositoryArtifact.ps1", action, StringComparison.Ordinal);
         Assert.Contains("POWERFORGE_BASELINE_ARTIFACT_RUN_ID", action, StringComparison.Ordinal);
+        Assert.Contains("POWERFORGE_DEPLOYMENT_JOB_ID", action, StringComparison.Ordinal);
+        Assert.Contains("deploymentJobId", action, StringComparison.Ordinal);
         Assert.Contains("POWERFORGE_GITHUB_API_URL", action, StringComparison.Ordinal);
         Assert.Contains("does not support hostname or base-path overrides", action, StringComparison.Ordinal);
         Assert.Contains("github-token: ${{ inputs.github-token }}", action, StringComparison.Ordinal);
@@ -561,156 +647,6 @@ public sealed class CloudflareIncrementalCachePurgeTests
         Assert.True(
             action.IndexOf("Purge changed Cloudflare URLs", StringComparison.Ordinal) <
             action.IndexOf("Upload deployed manifest baseline", StringComparison.Ordinal));
-    }
-
-    [Fact]
-    public void RepositoryArtifactLookup_ShouldSelectLatestNonExpiredArtifactAcrossBranches()
-    {
-        if (!CommandExists("pwsh"))
-            return;
-
-        var root = NewTempDirectory();
-        try
-        {
-            const string artifactName = "powerforge-cloudflare-manifest-v2-42-site";
-            var responsePath = Path.Combine(root, "response.json");
-            File.WriteAllText(responsePath,
-                $$"""
-                {
-                  "total_count": 4,
-                  "artifacts": [
-                    {
-                      "id": 4,
-                      "name": "different-artifact",
-                      "expired": false,
-                      "created_at": "2026-08-14T12:00:00Z",
-                      "workflow_run": { "id": 404, "head_branch": "main" }
-                    },
-                    {
-                      "id": 3,
-                      "name": "{{artifactName}}",
-                      "expired": true,
-                      "created_at": "2026-08-14T11:00:00Z",
-                      "workflow_run": { "id": 303, "head_branch": "main" }
-                    },
-                    {
-                      "id": 1,
-                      "name": "{{artifactName}}",
-                      "expired": false,
-                      "created_at": "2026-08-14T09:00:00Z",
-                      "workflow_run": { "id": 101, "head_branch": "feature/cache" }
-                    },
-                    {
-                      "id": 2,
-                      "name": "{{artifactName}}",
-                      "expired": false,
-                      "created_at": "2026-08-14T10:00:00Z",
-                      "workflow_run": { "id": 202, "head_branch": "main" }
-                    }
-                  ]
-                }
-                """);
-
-            var result = RunRepositoryArtifactLookup(root, responsePath, artifactName);
-
-            Assert.Equal("true", result["found"]);
-            Assert.Equal("202", result["run_id"]);
-            Assert.Equal("2", result["artifact_id"]);
-            Assert.Equal("2026-08-14T10:00:00.0000000+00:00", result["created_at"]);
-        }
-        finally
-        {
-            Directory.Delete(root, recursive: true);
-        }
-    }
-
-    [Fact]
-    public void BaselineOrder_ShouldDetectInterveningDeploymentAcrossRerunAttempts()
-    {
-        if (!CommandExists("pwsh")) return;
-        var root = NewTempDirectory();
-        try
-        {
-            var files = WriteDeploymentOrderFixture(root,
-                baselineRunId: 100, baselineAttempt: 1,
-                currentRunId: 100, currentAttempt: 2,
-                deployments:
-                [
-                    (30, 100, "2026-08-14T12:00:00Z"),
-                    (20, 200, "2026-08-14T11:00:00Z"),
-                    (10, 100, "2026-08-14T09:00:00Z")
-                ],
-                jobWindows:
-                [
-                    (100, 1, "2026-08-14T08:50:00Z", "2026-08-14T09:10:00Z"),
-                    (100, 2, "2026-08-14T11:50:00Z", "2026-08-14T12:10:00Z")
-                ]);
-
-            var result = RunBaselineOrder(root, files, 100, 2, 100);
-            Assert.Equal("false", result["stale"]);
-            Assert.Equal("false", result["use_previous"]);
-            Assert.Contains("intervening", result["reason"], StringComparison.OrdinalIgnoreCase);
-        }
-        finally { Directory.Delete(root, recursive: true); }
-    }
-
-    [Fact]
-    public void BaselineOrder_ShouldUseDeploymentHistoryEvenWhenInterveningPolicyNeverRecordedState()
-    {
-        if (!CommandExists("pwsh")) return;
-        var root = NewTempDirectory();
-        try
-        {
-            var files = WriteDeploymentOrderFixture(root,
-                100, 1, 300, 1,
-                [(30, 300, "2026-08-14T11:00:00Z"), (20, 200, "2026-08-14T10:00:00Z"), (10, 100, "2026-08-14T09:00:00Z")],
-                [(100, 1, "2026-08-14T08:50:00Z", "2026-08-14T09:10:00Z"), (300, 1, "2026-08-14T10:50:00Z", "2026-08-14T11:10:00Z")]);
-
-            var result = RunBaselineOrder(root, files, 300, 1, 100);
-            Assert.Equal("false", result["use_previous"]);
-            Assert.Contains("intervening", result["reason"], StringComparison.OrdinalIgnoreCase);
-        }
-        finally { Directory.Delete(root, recursive: true); }
-    }
-
-    [Fact]
-    public void BaselineOrder_ShouldUseExactBaselineAndSkipAStalePolicyJob()
-    {
-        if (!CommandExists("pwsh")) return;
-        var root = NewTempDirectory();
-        try
-        {
-            var files = WriteDeploymentOrderFixture(root,
-                100, 1, 300, 1,
-                [(30, 300, "2026-08-14T11:00:00Z"), (10, 100, "2026-08-14T09:00:00Z")],
-                [(100, 1, "2026-08-14T08:50:00Z", "2026-08-14T09:10:00Z"), (300, 1, "2026-08-14T10:50:00Z", "2026-08-14T11:10:00Z")]);
-            var ordered = RunBaselineOrder(root, files, 300, 1, 100);
-            Assert.Equal("true", ordered["use_previous"]);
-
-            AddDeployment(root, files, 40, 400, "2026-08-14T12:00:00Z");
-            var stale = RunBaselineOrder(root, files, 300, 1, 100);
-            Assert.Equal("true", stale["stale"]);
-            Assert.Equal("false", stale["use_previous"]);
-        }
-        finally { Directory.Delete(root, recursive: true); }
-    }
-
-    [Fact]
-    public void BaselineOrder_ShouldFailClosedUntilCurrentDeploymentIsIndexed()
-    {
-        if (!CommandExists("pwsh")) return;
-        var root = NewTempDirectory();
-        try
-        {
-            var files = WriteDeploymentOrderFixture(root,
-                100, 1, 300, 1,
-                [(10, 100, "2026-08-14T09:00:00Z")],
-                [(100, 1, "2026-08-14T08:50:00Z", "2026-08-14T09:10:00Z"), (300, 1, "2026-08-14T10:50:00Z", "2026-08-14T11:10:00Z")]);
-            var failure = RunBaselineOrderProcess(root, files, 300, 1, 100);
-            Assert.NotEqual(0, failure.ExitCode);
-            Assert.Contains("does not yet identify", failure.StandardError, StringComparison.OrdinalIgnoreCase);
-        }
-        finally { Directory.Delete(root, recursive: true); }
     }
 
     private static CloudflareDeploymentManifestEntry Entry(string path, char hashCharacter) => new()
@@ -786,189 +722,6 @@ public sealed class CloudflareIncrementalCachePurgeTests
         var root = Path.GetFullPath(Path.Combine(AppContext.BaseDirectory, "..", "..", "..", ".."));
         return Path.Combine([root, .. segments]);
     }
-
-    private static Dictionary<string, string> RunBaselineOrder(
-        string root,
-        DeploymentOrderFixture fixture,
-        long deploymentRunId,
-        int deploymentRunAttempt,
-        long baselineArtifactRunId)
-    {
-        var result = RunBaselineOrderProcess(root, fixture, deploymentRunId, deploymentRunAttempt, baselineArtifactRunId);
-        Assert.True(result.ExitCode == 0, $"Baseline-order validation failed ({result.ExitCode}). stdout: {result.StandardOutput} stderr: {result.StandardError}");
-
-        return File.ReadAllLines(result.OutputPath)
-            .Select(line => line.Split('=', 2))
-            .ToDictionary(parts => parts[0], parts => parts[1], StringComparer.Ordinal);
-    }
-
-    private static (int ExitCode, string StandardOutput, string StandardError, string OutputPath) RunBaselineOrderProcess(
-        string root,
-        DeploymentOrderFixture fixture,
-        long deploymentRunId,
-        int deploymentRunAttempt,
-        long baselineArtifactRunId)
-    {
-        var outputPath = Path.Combine(root, $"output-{Guid.NewGuid():N}.txt");
-        var wrapperPath = Path.Combine(root, $"deployment-wrapper-{Guid.NewGuid():N}.ps1");
-        File.WriteAllText(wrapperPath,
-            """
-            $ErrorActionPreference = 'Stop'
-            function global:Invoke-RestMethod {
-                param($Method, $Uri, $Headers)
-                if ($Uri -match '/actions/runs/([0-9]+)/attempts/([0-9]+)/jobs') {
-                    $path = Join-Path $env:POWERFORGE_TEST_ROOT "jobs-$($Matches[1])-$($Matches[2]).json"
-                } elseif ($Uri -match '/deployments/([0-9]+)/statuses') {
-                    $path = Join-Path $env:POWERFORGE_TEST_ROOT "statuses-$($Matches[1]).json"
-                } elseif ($Uri -match '/deployments\?') {
-                    $path = Join-Path $env:POWERFORGE_TEST_ROOT 'deployments.json'
-                } else {
-                    throw "Unexpected GitHub test URI: $Uri"
-                }
-                if (-not (Test-Path -LiteralPath $path -PathType Leaf)) { throw "Missing GitHub test response: $path" }
-                Get-Content -LiteralPath $path -Raw | ConvertFrom-Json
-            }
-            & $env:POWERFORGE_TEST_SCRIPT
-            """);
-        var startInfo = new ProcessStartInfo("pwsh")
-        {
-            WorkingDirectory = root,
-            RedirectStandardOutput = true,
-            RedirectStandardError = true,
-            UseShellExecute = false
-        };
-        startInfo.ArgumentList.Add("-NoLogo");
-        startInfo.ArgumentList.Add("-NoProfile");
-        startInfo.ArgumentList.Add("-File");
-        startInfo.ArgumentList.Add(wrapperPath);
-        startInfo.Environment["GITHUB_OUTPUT"] = outputPath;
-        startInfo.Environment["POWERFORGE_CLOUDFLARE_PREVIOUS_MANIFEST"] = fixture.PreviousManifest;
-        startInfo.Environment["POWERFORGE_CLOUDFLARE_BASELINE_STATE"] = fixture.BaselineState;
-        startInfo.Environment["POWERFORGE_BASELINE_ARTIFACT_RUN_ID"] = baselineArtifactRunId.ToString(System.Globalization.CultureInfo.InvariantCulture);
-        startInfo.Environment["POWERFORGE_GITHUB_API_URL"] = "https://api.github.test";
-        startInfo.Environment["POWERFORGE_GITHUB_REPOSITORY"] = "EvotecIT/Example";
-        startInfo.Environment["POWERFORGE_GITHUB_TOKEN"] = "test-token";
-        startInfo.Environment["POWERFORGE_DEPLOYMENT_RUN_ID"] = deploymentRunId.ToString(System.Globalization.CultureInfo.InvariantCulture);
-        startInfo.Environment["POWERFORGE_DEPLOYMENT_RUN_ATTEMPT"] = deploymentRunAttempt.ToString(System.Globalization.CultureInfo.InvariantCulture);
-        startInfo.Environment["POWERFORGE_TEST_ROOT"] = root;
-        startInfo.Environment["POWERFORGE_TEST_SCRIPT"] = RepoPath(".github", "actions", "powerforge-cloudflare-site-policy", "Resolve-PowerForgeCloudflareBaselineOrder.ps1");
-
-        using var process = Process.Start(startInfo) ?? throw new InvalidOperationException("Unable to start PowerShell baseline-order validation.");
-        var standardOutput = process.StandardOutput.ReadToEnd();
-        var standardError = process.StandardError.ReadToEnd();
-        process.WaitForExit();
-        return (process.ExitCode, standardOutput, standardError, outputPath);
-    }
-
-    private static Dictionary<string, string> RunRepositoryArtifactLookup(
-        string root,
-        string responsePath,
-        string artifactName)
-    {
-        var outputPath = Path.Combine(root, $"artifact-output-{Guid.NewGuid():N}.txt");
-        var wrapperPath = Path.Combine(root, $"artifact-wrapper-{Guid.NewGuid():N}.ps1");
-        File.WriteAllText(wrapperPath,
-            """
-            $ErrorActionPreference = 'Stop'
-            $global:PowerForgeTestArtifactResponse = Get-Content -LiteralPath $env:POWERFORGE_TEST_RESPONSE -Raw | ConvertFrom-Json
-            function global:Invoke-RestMethod {
-                param($Method, $Uri, $Headers)
-                return $global:PowerForgeTestArtifactResponse
-            }
-            & $env:POWERFORGE_TEST_SCRIPT
-            """);
-
-        var startInfo = new ProcessStartInfo("pwsh")
-        {
-            WorkingDirectory = root,
-            RedirectStandardOutput = true,
-            RedirectStandardError = true,
-            UseShellExecute = false
-        };
-        startInfo.ArgumentList.Add("-NoLogo");
-        startInfo.ArgumentList.Add("-NoProfile");
-        startInfo.ArgumentList.Add("-File");
-        startInfo.ArgumentList.Add(wrapperPath);
-        startInfo.Environment["GITHUB_OUTPUT"] = outputPath;
-        startInfo.Environment["POWERFORGE_ARTIFACT_NAME"] = artifactName;
-        startInfo.Environment["POWERFORGE_GITHUB_API_URL"] = "https://api.github.test";
-        startInfo.Environment["POWERFORGE_GITHUB_REPOSITORY"] = "EvotecIT/Example";
-        startInfo.Environment["POWERFORGE_GITHUB_TOKEN"] = "test-token";
-        startInfo.Environment["POWERFORGE_TEST_RESPONSE"] = responsePath;
-        startInfo.Environment["POWERFORGE_TEST_SCRIPT"] = RepoPath(".github", "actions", "powerforge-cloudflare-site-policy", "Resolve-PowerForgeRepositoryArtifact.ps1");
-
-        using var process = Process.Start(startInfo) ?? throw new InvalidOperationException("Unable to start repository-artifact validation.");
-        var standardOutput = process.StandardOutput.ReadToEnd();
-        var standardError = process.StandardError.ReadToEnd();
-        process.WaitForExit();
-        Assert.True(process.ExitCode == 0, $"Repository-artifact validation failed ({process.ExitCode}). stdout: {standardOutput} stderr: {standardError}");
-
-        return File.ReadAllLines(outputPath)
-            .Select(line => line.Split('=', 2))
-            .ToDictionary(parts => parts[0], parts => parts[1], StringComparer.Ordinal);
-    }
-
-    private static DeploymentOrderFixture WriteDeploymentOrderFixture(
-        string root,
-        long baselineRunId,
-        int baselineAttempt,
-        long currentRunId,
-        int currentAttempt,
-        (long Id, long RunId, string DeployedAt)[] deployments,
-        (long RunId, int Attempt, string StartedAt, string CompletedAt)[] jobWindows)
-    {
-        var previousManifest = Path.Combine(root, "manifest.json");
-        var baselineState = Path.Combine(root, "state.json");
-        File.WriteAllText(previousManifest, "{}");
-        File.WriteAllText(baselineState, $$"""{ "schemaVersion": 1, "deploymentRunId": "{{baselineRunId}}", "deploymentRunAttempt": {{baselineAttempt}} }""");
-        File.WriteAllText(Path.Combine(root, "deployments.json"), JsonSerializer.Serialize(
-            deployments.Select(item => new { id = item.Id, created_at = item.DeployedAt }).ToArray()));
-        foreach (var item in deployments)
-        {
-            var deployedAt = DateTimeOffset.Parse(item.DeployedAt, System.Globalization.CultureInfo.InvariantCulture);
-            var matchingJob = jobWindows.SingleOrDefault(job =>
-                job.RunId == item.RunId &&
-                deployedAt >= DateTimeOffset.Parse(job.StartedAt, System.Globalization.CultureInfo.InvariantCulture) &&
-                deployedAt <= DateTimeOffset.Parse(job.CompletedAt, System.Globalization.CultureInfo.InvariantCulture));
-            var jobId = matchingJob == default ? item.Id : matchingJob.RunId * 10 + matchingJob.Attempt;
-            File.WriteAllText(Path.Combine(root, $"statuses-{item.Id}.json"), JsonSerializer.Serialize(new[]
-            {
-                new { state = "success", environment_url = "https://example.test/", log_url = $"https://github.test/EvotecIT/Example/actions/runs/{item.RunId}/job/{jobId}", created_at = item.DeployedAt }
-            }));
-        }
-        foreach (var job in jobWindows)
-        {
-            File.WriteAllText(Path.Combine(root, $"jobs-{job.RunId}-{job.Attempt}.json"), JsonSerializer.Serialize(new
-            {
-                total_count = 1,
-                jobs = new[]
-                {
-                    new
-                    {
-                        id = job.RunId * 10 + job.Attempt,
-                        started_at = job.StartedAt,
-                        completed_at = job.CompletedAt,
-                        steps = new[] { new { name = "Deploy to GitHub Pages", conclusion = "success" } }
-                    }
-                }
-            }));
-        }
-        return new DeploymentOrderFixture(previousManifest, baselineState);
-    }
-
-    private static void AddDeployment(string root, DeploymentOrderFixture fixture, long id, long runId, string deployedAt)
-    {
-        var path = Path.Combine(root, "deployments.json");
-        var deployments = JsonNode.Parse(File.ReadAllText(path))!.AsArray();
-        deployments.Insert(0, new JsonObject { ["id"] = id, ["created_at"] = deployedAt });
-        File.WriteAllText(path, deployments.ToJsonString());
-        File.WriteAllText(Path.Combine(root, $"statuses-{id}.json"), JsonSerializer.Serialize(new[]
-        {
-            new { state = "success", environment_url = "https://example.test/", log_url = $"https://github.test/EvotecIT/Example/actions/runs/{runId}/job/{id}", created_at = deployedAt }
-        }));
-    }
-
-    private sealed record DeploymentOrderFixture(string PreviousManifest, string BaselineState);
 
     private static bool CommandExists(string command)
     {
